add redirect to people page
