--- src/components/JEDI/index.js
@@ -1,3 +1,0 @@
-import JEDI from './JEDI';
-
-export default JEDI;
