+++ src/components/People/People.jsx
@@ -1,9 +1,43 @@
-import React from 'react';
+
+import React, { useState } from 'react';
+import Table from '../common/Table';
+import Form from '../common/Form';
+
+const data = [
+  { first: 'Mark', last: 'Otto', handle: '@motto', id: '1' },
+  { first: 'Carl', last: 'Reno', handle: '@ceno', id: '2' },
+  { first: 'Steve', last: 'Smith', handle: '@ssteve', id: '3' }
+]
+
+const columns = Object.keys(data[0]);
 
 const People = () => {
+  const [people, setPeople] = useState(data);
+
+  const handleAppPerson = (personData) => {
+    const data = [...people, personData];
+    setPeople(data)
+  }
+
+  const getInitialPeopleData = () => {
+    return columns.reduce((cols, columnName) => {
+      cols[columnName] = "";
+      return cols;
+    }, {})
+  }
+
   return (
-    <div>
-      People
+    <div className="container">
+      <Table
+        data={people}
+        columns={columns}
+        tableDescriptor="People"
+      />
+      <Form
+        initialData={getInitialPeopleData()}
+        columns={columns}
+        onAddData={handleAppPerson}
+      />
     </div>
   )
 };
